Support custom default mode mappings
Without this it is impossible to set default mode mappings as it forces it to
the default ones from core.

This makes it possible to specify the store to use in config. The following
example shows how you would switch from using `default_application` to
`somestore` for MODE_APPLICATION stores.
```
$CFG->tool_forcedcache_config_array = [
  'stores' => [
    'somestore' => [ ... ],
  ],
  'modemappings' => [
    'application' => 'somestore',
  ],
];
```

This functionality is useful when using $CFG->enable_read_only_sessions - as
adhoc session stores would otherwise store data in the default_session store -
which triggers an immediate error for the user they can't escape in M3.11+
without killing their session.